Feat: upload-zzal 경로 prefetch 적용
